+++ src/pages/index.js
@@ -7,4 +7,4 @@ export * from "./Users";
 export * from "./Chat";
 export * from "./Dashboard";
 export * from "./TermsAndConditions";
-export * from "./CallStatsChart"
+export * from "./CallStatsPage"
